[DOCS-3364] Remove `null` fields from Customer response
