+++ src/main/java/fauna/sample/controllers/products/Product.java
@@ -1,5 +1,6 @@
 package fauna.sample.controllers.products;
 
+import com.fauna.annotation.FaunaField;
 import com.fauna.annotation.FaunaId;
 
 public class Product {
@@ -39,7 +40,10 @@ public class Product {
     private String name;
     private String description;
     private Integer price;
-    private Integer stock;
+    // Fauna object and document fields will be matched to class fields by default, but
+    // you can override the name with @FaunaField( name = "my_name" )
+    @FaunaField( name = "stock")
+    private Integer stockQuantity;
     private Category category;
 
     public String getId() {
@@ -58,8 +62,8 @@ public class Product {
         return price;
     }
 
-    public Integer getStock() {
-        return stock;
+    public Integer getStockQuantity() {
+        return stockQuantity;
     }
 
     public Category getCategory() {
