+++ src/main/java/fauna/sample/controllers/products/ProductRequest.java
@@ -1,10 +1,15 @@
 package fauna.sample.controllers.products;
 
+import com.fauna.annotation.FaunaField;
+
 public class ProductRequest {
     private String name;
     private String description;
     private Integer price;
-    private Integer stock;
+    // Fauna object and document fields will be matched to class fields by default, but
+    // you can override the name with @FaunaField( name = "my_name" )
+    @FaunaField( name = "stock")
+    private Integer stockQuantity;
     private String category;
 
     public String getName() {
@@ -19,8 +24,8 @@ public class ProductRequest {
         return price;
     }
 
-    public Integer getStock() {
-        return stock;
+    public Integer getStockQuantity() {
+        return stockQuantity;
     }
 
     public String getCategory() {
